remove additional test for News Store
